requirements.txt
fdfs_client
asgiref==3.2.5
Django==3.0.4
django-discover-runner==1.0
django-role-permissions==2.2.1
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
